save searchresult into a localstorage
